+++ pymatgen/analysis/elasticity/tests/test_tensors.py
@@ -223,10 +223,11 @@ class TensorBaseTest(PymatgenTest):
         self.assertTrue(self.fit_r4.is_fit_to_structure(self.structure))
 
     def test_convert_to_ieee(self):
-        for xtal in self.ieee_data.keys():
-            orig = TensorBase(self.ieee_data[xtal]['original_tensor'])
-            ieee = TensorBase(self.ieee_data[xtal]['ieee_tensor'])
-            struct = Structure.from_dict(self.ieee_data[xtal]['structure'])
+        for entry in self.ieee_data:
+            xtal = entry['xtal']
+            orig = TensorBase(entry['original_tensor'])
+            ieee = TensorBase(entry['ieee_tensor'])
+            struct = Structure.from_dict(entry['structure'])
             diff = np.max(abs(ieee - orig.convert_to_ieee(struct)))
             err_msg = "{} IEEE conversion failed with max diff {}".format(
                 xtal, diff)